sw: port the FreeRTOS demo to rv64, retiring the D13 deferral
D13 froze the FreeRTOS demo at rv32 because its FROST-local port
needed a full sd/ld context-frame rewrite; that rewrite became a
solved, six-times-practiced pattern during the hardware bring-up's
trap-frame wave, so the deferral is retired early rather than at
phase exit:

- port_frost_asm.S: STORE/LOAD/XB per-XLEN macros across the whole
  31-slot context frame (portCONTEXT_SIZE 128 -> 256 at rv64),
  XLEN-wide accesses for the pxCurrentTCB/pxTopOfStack/
  uxCriticalNesting pointer traffic, and the timer-mcause literal
  with the interrupt bit at XLEN-1.
- portmacro.h: StackType_t/BaseType_t/UBaseType_t/pointer-size types
  follow the register width; TickType_t stays 32-bit (atomic at both).
- FreeRTOSConfig.h: the heap doubles at rv64 -- task stacks are sized
  in StackType_t words, so their byte footprint doubles and the old
  8 KiB heap would have failed xTaskCreate silently.
- pxPortInitialiseStack needed nothing: it builds the frame in
  StackType_t cells, so it scaled with the type change.

Validated: <<PASS>> on rv64 X3 silicon plus both XLENs in sim. The
registry twin generates again (exclusion removed), and the phase
plan's D13 entry carries the as-built note.

+++ sw/apps/freertos_demo/port_frost.c
@@ -145,7 +145,7 @@ extern void *volatile pxCurrentTCB;
 void vPortDebugTCB(char marker)
 {
     UART_TX = marker;
-    print_hex((uint32_t) pxCurrentTCB);
+    print_hex((uint32_t) (uintptr_t) pxCurrentTCB);
 }
 
 /* Debug: print RA value */
